Start of Media Content site
This site is for a project in an HTML course. The concept is assuming
all data (info from sites like Hulu, Netflix, Twitter etc) is provided
to the front end developer, display this content to the user in a
meaningful way so that they can find out more about their media
consumption.
So far a main page / home page has been created with just three buttons
which will eventually be the main way a user can select one of three
tabs to see information about their media habits - using js and ajax to
hide or show one of the three sections